refactor(jit): move expression parser to runtime
