+++ packages/jit-html-jsdom/src/index.ts
@@ -1,7 +1,6 @@
 import {
   DefaultBindingLanguage as JitDefaultBindingLanguage,
   DefaultBindingSyntax as JitDefaultBindingSyntax,
-  DefaultComponents as JitDefaultComponents
 } from '@aurelia/jit';
 import {
   DefaultBindingLanguage as JitHtmlDefaultBindingLanguage,
@@ -31,7 +30,6 @@ export const JitHtmlJsdomConfiguration = {
       .register(
         ...JitDefaultBindingLanguage,
         ...JitDefaultBindingSyntax,
-        ...JitDefaultComponents,
         ...JitHtmlDefaultBindingLanguage,
         ...JitHtmlDefaultComponents
       );
